Load generic FFT fallbacks in FourierTransforms only optionally
