+++ src/DFTK.jl
@@ -158,6 +158,16 @@ function __init__()
     @require IterativeSolvers="42fd0dbc-a981-5370-80f2-aaf504508153" begin
         include("eigen/diag_lobpcg_itsolve.jl")
     end
+
+    # Load generic FFT stuff once IntervalArithmetic or DoubleFloats are used
+    # The global variable GENERIC_FFT_LOADED makes sure that things are only
+    # included once.
+    @require IntervalArithmetic="d1acc4aa-44c8-5952-acd4-ba5d80a2a253" begin
+        !isdefined(DFTK, :GENERIC_FFT_LOADED) && include("fft_generic.jl")
+    end
+    @require DoubleFloats="497a8b3b-efae-58df-a0af-a86822472b78" begin
+        !isdefined(DFTK, :GENERIC_FFT_LOADED) && include("fft_generic.jl")
+    end
 end
 
 end # module DFTK
